renamed project (hotcache was taken); made it compatible with go get/install; added license;
