+++ scanner/scan.go
